+++ CHANGELOG.md
@@ -57,6 +57,7 @@
         * Change: Remove 'test-' prefix from unit test files
         * Change: Rename `phpunit.xml` to `phpunit.xml.dist` per best practices
     * New: Add additional punctuation-related test cases
+    * New: Add helper function `get_core_filters()` to DRY up data reuse
 * New: Add a few more possible TODO items
 
 ## 4.0 _(2020-07-16)_